fix eri for hybrid case, add random int to testid
