+++ scripts/create_test
@@ -16,7 +16,7 @@ from CIME.utils import expect, convert_to_seconds, compute_total_time, convert_t
 from CIME.XML.machines import Machines
 from CIME.case import Case
 
-import argparse, math, glob
+import argparse, math, glob, random
 
 logger = logging.getLogger(__name__)
 
@@ -257,7 +257,7 @@ OR
         args.no_batch = True
 
     if args.test_id is None:
-        args.test_id = CIME.utils.get_timestamp()
+        args.test_id = "%s.%s"%(CIME.utils.get_timestamp(), random.randrange(0,99,1))
 
     if args.testfile is not None:
         with open(args.testfile, "r") as fd:
